requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
kagglehub
tensorboard
gradio
Pillow
tqdm
matplotlib
safetensors
huggingface_hub